requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
anthropic>=0.42.0
deepgram-sdk>=3.7.0
elevenlabs>=1.9.0
httpx>=0.27.0
